Moved isFolderEmpty into folderManagement
