+++ __tests__/backend/move/folderManagement.test.ts
@@ -1,5 +1,6 @@
 import {
   deleteFolderIfEmpty,
+  isFolderEmpty,
   listFilesInFolder,
   listFoldersInFolder,
 } from "../../../src/backend/move/folderManagement";
@@ -126,6 +127,74 @@ test("listFoldersInFolder works correctly", () => {
   expect(list.mock.calls[0][0].fields).toContain("id");
   expect(list.mock.calls[0][0].fields).toContain("title");
   expect(list.mock.calls[0][0].fields).toContain("canMoveItemOutOfDrive");
+});
+
+test("isFolderEmpty works correctly with an empty folder", () => {
+  interface File {
+    id?: string;
+  }
+  interface FileList {
+    items?: Array<File>;
+  }
+  interface ListFilesOptions {
+    q?: string;
+    includeItemsFromAllDrives?: boolean;
+    supportsAllDrives?: boolean;
+    maxResults?: number;
+    fields?: string;
+  }
+
+  const rawResponse = {
+    items: [],
+  };
+  const list = jest
+    .fn<FileList, [optionalArgs: ListFilesOptions]>()
+    .mockReturnValueOnce(rawResponse);
+  global.Drive = {
+    Files: {
+      list,
+    },
+  };
+
+  expect(isFolderEmpty("ID_FOLDER")).toBe(true);
+  expect(list.mock.calls.length).toBe(1);
+  expect(list.mock.calls[0][0].q).toContain("ID_FOLDER");
+  expect(list.mock.calls[0][0].includeItemsFromAllDrives).toBe(true);
+  expect(list.mock.calls[0][0].supportsAllDrives).toBe(true);
+});
+
+test("isFolderEmpty works correctly with a non-empty folder", () => {
+  interface File {
+    id?: string;
+  }
+  interface FileList {
+    items?: Array<File>;
+  }
+  interface ListFilesOptions {
+    q?: string;
+    includeItemsFromAllDrives?: boolean;
+    supportsAllDrives?: boolean;
+    maxResults?: number;
+    fields?: string;
+  }
+
+  const rawResponse = {
+    items: [{ id: "ID_FILE" }],
+  };
+  const list = jest
+    .fn<FileList, [optionalArgs: ListFilesOptions]>()
+    .mockReturnValueOnce(rawResponse);
+  global.Drive = {
+    Files: {
+      list,
+    },
+  };
+
+  expect(isFolderEmpty("ID_FOLDER")).toBe(false);
+  expect(list.mock.calls.length).toBe(1);
+  expect(list.mock.calls[0][0].q).toContain("ID_FOLDER");
+  expect(list.mock.calls[0][0].includeItemsFromAllDrives).toBe(true);
+  expect(list.mock.calls[0][0].supportsAllDrives).toBe(true);
 });
 
 test.each(["owner", "organizer"] as Array<
