--- src/backend/move/isFolderEmpty.ts
@@ -1,10 +1,0 @@
-export function isFolderEmpty(folderID: string): boolean {
-  const response = Drive.Files!.list({
-    q: '"' + folderID + '" in parents and trashed = false',
-    includeItemsFromAllDrives: true,
-    supportsAllDrives: true,
-    maxResults: 1,
-    fields: "items(id)",
-  });
-  return response.items!.length === 0;
-}
